breakout user api tests
