+++ tdrs-backend/tdpservice/users/test/test_api/test_set_profile.py
@@ -1,10 +1,9 @@
-"""API User Tests."""
+"""API User Set Profile Tests."""
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group, Permission
 from django.core.management import call_command
 import pytest
 from rest_framework import status
-from ...stts.models import STT
+from tdpservice.stts.models import STT
 
 User = get_user_model()
 
@@ -14,40 +13,6 @@ User = get_user_model()
 def create_test_users():
     """Create users for each group."""
     call_command("generate_test_users")
-
-
-@pytest.mark.django_db
-def test_retrieve_user(api_client, user):
-    """Test user retrieval."""
-    api_client.login(username=user.username, password="test_password")
-    response = api_client.get(f"/v1/users/{user.pk}/")
-    assert response.status_code == status.HTTP_200_OK
-    assert response.data["username"] == user.username
-
-
-@pytest.mark.django_db
-def test_can_update_own_user(api_client, user):
-    """Test a user can update their own user."""
-    api_client.login(username=user.username, password="test_password")
-    response = api_client.patch(f"/v1/users/{user.pk}/", {"first_name": "Jane"})
-    assert response.status_code == status.HTTP_200_OK
-    assert response.data["first_name"] == "Jane"
-    assert User.objects.filter(first_name="Jane").exists()
-
-
-@pytest.mark.django_db
-def test_cannot_update_user_anonymously(api_client, user):
-    """Test an unauthenticated user cannot update a user."""
-    response = api_client.patch(f"/v1/users/{user.pk}/", {"first_name": "Jane"})
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_create_user(api_client, user_data):
-    """Test user creation."""
-    response = api_client.post("/v1/users/", user_data)
-    assert response.status_code == status.HTTP_201_CREATED
-    assert User.objects.filter(username=user_data["username"]).exists()
 
 
 @pytest.mark.django_db
@@ -349,183 +314,3 @@ def test_set_profile_data_missing_first_name_field(api_client, user):
     api_client.login(username=user.username, password="test_password")
     response = api_client.patch("/v1/users/set_profile/", {"last_name": "Heather", },)
     assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
-@pytest.mark.django_db
-def test_role_list(api_client, create_test_users):
-    """Test role list."""
-    # Groups are populated in a data migrations, so are already available.
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.get("/v1/roles/")
-    assert response.status_code == status.HTTP_200_OK
-    role_names = {group["name"] for group in response.data}
-    assert role_names == {"OFA Admin", "OFA Analyst", "Data Prepper"}
-
-
-@pytest.mark.django_db
-def test_role_list_unauthorized(api_client, create_test_users):
-    """Data prepper does not have access."""
-    # Groups are populated in a data migrations, so are already available.
-    api_client.login(username="test__data_prepper", password="test_password")
-    response = api_client.get("/v1/roles/")
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_role_create_forbidden(api_client, create_test_users):
-    """Test creating a role is no longer allowed."""
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.post("/v1/roles/", {"name": "Test Role"})
-    assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
-
-
-@pytest.mark.django_db
-def test_role_create_unauthorized(api_client, create_test_users):
-    """Test data prepper does not have access."""
-    api_client.login(username="test__data_prepper", password="test_password")
-    response = api_client.post("/v1/roles/", {"name": "Test Role"})
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_role_create_with_permission_forbidden(api_client, create_test_users):
-    """Test creating a role with a permission is no longer allowed."""
-    permission = Permission.objects.first()
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.post(
-        "/v1/roles/", {"name": "Test Role", "permissions": [permission.id]}
-    )
-    assert response.status_code == status.HTTP_405_METHOD_NOT_ALLOWED
-
-
-@pytest.mark.django_db
-def test_role_update_not_found(api_client, create_test_users):
-    """Test role update no longer exists."""
-    group = Group.objects.get(name="OFA Admin")
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.patch(f"/v1/roles/{group.id}/", {"name": "staff"})
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_role_update_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     group = Group.objects.get(name="Data Prepper")
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.patch(f"/v1/roles/{group.id}/", {"name": "staff"})
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_role_delete_not_found(api_client, create_test_users):
-    """Test role deletion no longer exists."""
-    group = Group.objects.get(name="OFA Admin")
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.delete(f"/v1/roles/{group.id}/")
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_role_delete_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     group = Group.objects.get(name="Data Prepper")
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.delete(f"/v1/roles/{group.id}/")
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_permission_list_not_found(api_client, create_test_users):
-    """Test permission list no longer exists."""
-    # Django comes with some permissions, so we'll just test those.
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.get("/v1/permissions/")
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_permission_list_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     # Django comes with some permissions, so we'll just test those.
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.get("/v1/permissions/")
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_permission_create_not_found(api_client, create_test_users):
-    """Test permission creation no longer exists."""
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.post(
-        "/v1/permissions/", {"codename": "foo", "name": "Foo", "content_type": None}
-    )
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_permission_create_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.post(
-#         "/v1/permissions/", {"codename": "foo", "name": "Foo", "content_type": None}
-#     )
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-# @pytest.mark.django_db
-# def test_permission_create_with_content_type(api_client, create_test_users):
-#     """Test can create a permission with a content type."""
-#     content_type = ContentType.objects.get_for_model(User)
-#     api_client.login(username="test__ofa_admin", password="test_password")
-#     response = api_client.post(
-#         "/v1/permissions/",
-#         {
-#             "codename": "foo",
-#             "name": "Foo",
-#             "content_type": f"{content_type.app_label}.{content_type.model}",
-#         },
-#     )
-#     assert response.status_code == status.HTTP_201_CREATED
-#     assert response.data["codename"] == "foo"
-#     assert response.data["name"] == "Foo"
-#     assert Permission.objects.filter(codename="foo").exists()
-
-
-@pytest.mark.django_db
-def test_permission_update_not_found(api_client, create_test_users):
-    """Test permission update."""
-    permission = Permission.objects.first()
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.patch(
-        f"/v1/permissions/{permission.id}/", {"codename": "foo"}
-    )
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_permission_update_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     permission = Permission.objects.first()
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.patch(
-#         f"/v1/permissions/{permission.id}/", {"codename": "foo"}
-#     )
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-@pytest.mark.django_db
-def test_permission_delete_not_found(api_client, create_test_users):
-    """Test permission deletion no longer exists."""
-    permission = Permission.objects.first()
-    api_client.login(username="test__ofa_admin", password="test_password")
-    response = api_client.delete(f"/v1/permissions/{permission.id}/")
-    assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# @pytest.mark.django_db
-# def test_permission_delete_unauthorized(api_client, create_test_users):
-#     """Test data prepper does not have access."""
-#     permission = Permission.objects.first()
-#     api_client.login(username="test__data_prepper", password="test_password")
-#     response = api_client.delete(f"/v1/permissions/{permission.id}/")
-#     assert response.status_code == status.HTTP_403_FORBIDDEN
